requirements.txt
Flask==2.3.2
numpy==1.24.4
Pillow==10.2.0
scikit-learn==1.3.2